+++ components/common/demo/components/tree/tree.vue
@@ -135,5 +135,8 @@ export default {
 };
 </script>
 
-<style lang="less" scoped>
+<style lang="less">
+    .node-wrapper .panel ul {
+        margin-bottom: 0;
+    }
 </style>